refactor network reducer constants and unused reducer
When first writing the laboratory, we were not sure what was best to put
in the reducer. After a bit more experience, it is more clear that we
shouldn't be storing static data in the reducer since it clutters up the
reducers and can be done in a cleaner way (by having a constants file).

The presence of a potentially dynamic available networks can mess up how
we serialize and deserialize the network part of the state. I do not
foresee having an option in the laboratory to add new horizon servers
any time soon and this is a good tradeoff to make for safety.