Switch to synchronous communication with Ledger device
The newest version of Stellar app for Ledger locks at each function
call. It means that it's not possible to call two different functions at
the same time, it results in an error:

```
Ledger Device is busy (lock getPublicKey)
```

This commit changes the `Promise.all` call to the sequential code.